requirements.txt
websocket-client==1.6.4
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
streamlit==1.31.0
plotly==5.18.0
pandas==2.2.0
pytest==7.4.0
pytest-cov==4.1.0
pre-commit==3.5.0
black==23.7.0
flake8==6.1.0
isort==5.12.0
mypy==1.4.1